+++ axum-extra/src/extract/cookie/mod.rs
@@ -168,7 +168,7 @@ impl CookieJar {
     /// use axum::response::IntoResponse;
     ///
     /// async fn handle(jar: CookieJar) -> CookieJar {
-    ///     jar.remove(Cookie::named("foo"))
+    ///     jar.remove(Cookie::from("foo"))
     /// }
     /// ```
     #[must_use]
@@ -234,6 +234,7 @@ fn set_cookies(jar: cookie::CookieJar, headers: &mut HeaderMap) {
 mod tests {
     use super::*;
     use axum::{body::Body, extract::FromRef, http::Request, routing::get, Router};
+    use http_body_util::BodyExt;
     use tower::ServiceExt;
 
     macro_rules! cookie_test {
@@ -249,7 +250,7 @@ mod tests {
                 }
 
                 async fn remove_cookie(jar: $jar) -> impl IntoResponse {
-                    jar.remove(Cookie::named("key"))
+                    jar.remove(Cookie::from("key"))
                 }
 
                 let state = AppState {
@@ -378,7 +379,7 @@ mod tests {
         B: axum::body::HttpBody,
         B::Error: std::fmt::Debug,
     {
-        let bytes = hyper::body::to_bytes(body).await.unwrap();
+        let bytes = body.collect().await.unwrap().to_bytes();
         String::from_utf8(bytes.to_vec()).unwrap()
     }
 }