+++ axum/src/extension.rs
@@ -98,7 +98,7 @@ axum_core::__impl_deref!(Extension);
 
 impl<T> IntoResponseParts for Extension<T>
 where
-    T: Send + Sync + 'static,
+    T: Clone + Send + Sync + 'static,
 {
     type Error = Infallible;
 
@@ -110,7 +110,7 @@ where
 
 impl<T> IntoResponse for Extension<T>
 where
-    T: Send + Sync + 'static,
+    T: Clone + Send + Sync + 'static,
 {
     fn into_response(self) -> Response {
         let mut res = ().into_response();