+++ axum-extra/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning].
 
 # Unreleased
 
+- **added:** `OptionalQuery` extractor ([#2310])
 - **added:** `TypedHeader` which used to be in `axum` ([#1850])
 - **added:** `Clone` implementation for `ErasedJson` ([#2142])
 - **breaking:** Update to prost 0.12. Used for the `Protobuf` extractor
@@ -14,6 +15,7 @@ and this project adheres to [Semantic Versioning].
 
 [#1850]: https://github.com/tokio-rs/axum/pull/1850
 [#2142]: https://github.com/tokio-rs/axum/pull/2142
+[#2310]: https://github.com/tokio-rs/axum/pull/2310 
 
 # 0.7.4 (18. April, 2023)